+++ CHANGELOG.md
@@ -200,6 +200,7 @@ here is a non-exhaustive list of significant departures from the original gem:
 * Fix {OpenHAB::DSL::Items::Ensure ensure} to work with conversions-from-string that are handled by openhab-core.
 * Avoid stack overflow issues when all of ActiveSupport is required.
 * Don't swallow exceptions inside of `rule` blocks - just let them propagate up.
+* Fix changed duration when only the `from` state is given
 
 ## [4.45.2](https://github.com/boc-tothefuture/openhab-jruby/compare/4.45.1...4.45.2) (2022-10-02)